[feature] : implement send all data of admin to client

+++ lib/features/admin/domain/models/admin.dart
@@ -47,7 +47,7 @@ class Admin {
   ///
   /// name of admin
   ///
-  short? phone;
+  int? phone;
 
   ///
   /// name of admin
